+++ include/extras-kcolorpicker/libKColorPicker.h
@@ -1401,7 +1401,7 @@ void k_colorpicker__kcolorpicker_set_graphics_effect(void* self, void* effect);
 /// @param self kColorPicker__KColorPicker*
 /// @param type enum Qt__GestureType
 ///
-void k_colorpicker__kcolorpicker_grab_gesture(void* self, int64_t type);
+void k_colorpicker__kcolorpicker_grab_gesture(void* self, int32_t type);
 
 /// Inherited from QWidget
 ///
@@ -1410,7 +1410,7 @@ void k_colorpicker__kcolorpicker_grab_gesture(void* self, int64_t type);
 /// @param self kColorPicker__KColorPicker*
 /// @param type enum Qt__GestureType
 ///
-void k_colorpicker__kcolorpicker_ungrab_gesture(void* self, int64_t type);
+void k_colorpicker__kcolorpicker_ungrab_gesture(void* self, int32_t type);
 
 /// Inherited from QWidget
 ///
@@ -2399,7 +2399,7 @@ void k_colorpicker__kcolorpicker_set_parent(void* self, void* parent);
 /// @param parent QWidget*
 /// @param f flag of enum Qt__WindowType
 ///
-void k_colorpicker__kcolorpicker_set_parent2(void* self, void* parent, int64_t f);
+void k_colorpicker__kcolorpicker_set_parent2(void* self, void* parent, int32_t f);
 
 /// Inherited from QWidget
 ///
@@ -2477,9 +2477,9 @@ void k_colorpicker__kcolorpicker_add_action(void* self, void* action);
 /// [Upstream resources](https://doc.qt.io/qt-6/qwidget.html#addActions)
 ///
 /// @param self kColorPicker__KColorPicker*
-/// @param actions libqt_list /* of QAction* */
+/// @param actions libqt_list of QAction*
 ///
-void k_colorpicker__kcolorpicker_add_actions(void* self, libqt_list actions);
+void k_colorpicker__kcolorpicker_add_actions(void* self, libqt_list /* of QAction* */ actions);
 
 /// Inherited from QWidget
 ///
@@ -2487,9 +2487,9 @@ void k_colorpicker__kcolorpicker_add_actions(void* self, libqt_list actions);
 ///
 /// @param self kColorPicker__KColorPicker*
 /// @param before QAction*
-/// @param actions libqt_list /* of QAction* */
+/// @param actions libqt_list of QAction*
 ///
-void k_colorpicker__kcolorpicker_insert_actions(void* self, void* before, libqt_list actions);
+void k_colorpicker__kcolorpicker_insert_actions(void* self, void* before, libqt_list /* of QAction* */ actions);
 
 /// Inherited from QWidget
 ///
@@ -2516,7 +2516,9 @@ void k_colorpicker__kcolorpicker_remove_action(void* self, void* action);
 ///
 /// @param self kColorPicker__KColorPicker*
 ///
-libqt_list /* of QAction* */ k_colorpicker__kcolorpicker_actions(void* self);
+/// @return libqt_list of QAction*
+///
+libqt_list k_colorpicker__kcolorpicker_actions(void* self);
 
 /// Inherited from QWidget
 ///
@@ -2573,7 +2575,7 @@ QWidget* k_colorpicker__kcolorpicker_parent_widget(void* self);
 /// @param self kColorPicker__KColorPicker*
 /// @param type flag of enum Qt__WindowType
 ///
-void k_colorpicker__kcolorpicker_set_window_flags(void* self, int64_t type);
+void k_colorpicker__kcolorpicker_set_window_flags(void* self, int32_t type);
 
 /// Inherited from QWidget
 ///
@@ -2583,7 +2585,7 @@ void k_colorpicker__kcolorpicker_set_window_flags(void* self, int64_t type);
 ///
 /// @return flag of enum Qt__WindowType
 ///
-int64_t k_colorpicker__kcolorpicker_window_flags(void* self);
+int32_t k_colorpicker__kcolorpicker_window_flags(void* self);
 
 /// Inherited from QWidget
 ///
@@ -2592,7 +2594,7 @@ int64_t k_colorpicker__kcolorpicker_window_flags(void* self);
 /// @param self kColorPicker__KColorPicker*
 /// @param param1 enum Qt__WindowType
 ///
-void k_colorpicker__kcolorpicker_set_window_flag(void* self, int64_t param1);
+void k_colorpicker__kcolorpicker_set_window_flag(void* self, int32_t param1);
 
 /// Inherited from QWidget
 ///
@@ -2601,7 +2603,7 @@ void k_colorpicker__kcolorpicker_set_window_flag(void* self, int64_t param1);
 /// @param self kColorPicker__KColorPicker*
 /// @param type flag of enum Qt__WindowType
 ///
-void k_colorpicker__kcolorpicker_override_window_flags(void* self, int64_t type);
+void k_colorpicker__kcolorpicker_override_window_flags(void* self, int32_t type);
 
 /// Inherited from QWidget
 ///
@@ -2611,7 +2613,7 @@ void k_colorpicker__kcolorpicker_override_window_flags(void* self, int64_t type)
 ///
 /// @return enum Qt__WindowType
 ///
-int64_t k_colorpicker__kcolorpicker_window_type(void* self);
+int32_t k_colorpicker__kcolorpicker_window_type(void* self);
 
 /// Inherited from QWidget
 ///
@@ -2822,7 +2824,7 @@ void k_colorpicker__kcolorpicker_on_custom_context_menu_requested(void* self, vo
 ///
 /// @return flag of enum Qt__InputMethodHint
 ///
-int64_t k_colorpicker__kcolorpicker_input_method_hints(void* self);
+int32_t k_colorpicker__kcolorpicker_input_method_hints(void* self);
 
 /// Inherited from QWidget
 ///
@@ -2831,7 +2833,7 @@ int64_t k_colorpicker__kcolorpicker_input_method_hints(void* self);
 /// @param self kColorPicker__KColorPicker*
 /// @param hints flag of enum Qt__InputMethodHint
 ///
-void k_colorpicker__kcolorpicker_set_input_method_hints(void* self, int64_t hints);
+void k_colorpicker__kcolorpicker_set_input_method_hints(void* self, int32_t hints);
 
 /// Inherited from QWidget
 ///
@@ -2916,7 +2918,7 @@ QPixmap* k_colorpicker__kcolorpicker_grab1(void* self, void* rectangle);
 /// @param type enum Qt__GestureType
 /// @param flags flag of enum Qt__GestureFlag
 ///
-void k_colorpicker__kcolorpicker_grab_gesture2(void* self, int64_t type, int32_t flags);
+void k_colorpicker__kcolorpicker_grab_gesture2(void* self, int32_t type, int32_t flags);
 
 /// Inherited from QWidget
 ///
@@ -2956,7 +2958,7 @@ void k_colorpicker__kcolorpicker_set_shortcut_auto_repeat2(void* self, int id, b
 /// @param param1 enum Qt__WindowType
 /// @param on bool
 ///
-void k_colorpicker__kcolorpicker_set_window_flag2(void* self, int64_t param1, bool on);
+void k_colorpicker__kcolorpicker_set_window_flag2(void* self, int32_t param1, bool on);
 
 /// Inherited from QWidget
 ///
@@ -2985,7 +2987,7 @@ QWidget* k_colorpicker__kcolorpicker_create_window_container2(void* window, void
 /// @param parent QWidget*
 /// @param flags flag of enum Qt__WindowType
 ///
-QWidget* k_colorpicker__kcolorpicker_create_window_container3(void* window, void* parent, int64_t flags);
+QWidget* k_colorpicker__kcolorpicker_create_window_container3(void* window, void* parent, int32_t flags);
 
 /// Inherited from QObject
 ///
@@ -3097,7 +3099,9 @@ void k_colorpicker__kcolorpicker_kill_timer2(void* self, int32_t id);
 ///
 /// @param self kColorPicker__KColorPicker*
 ///
-libqt_list /* of QObject* */ k_colorpicker__kcolorpicker_children(void* self);
+/// @return libqt_list of QObject*
+///
+libqt_list k_colorpicker__kcolorpicker_children(void* self);
 
 /// Inherited from QObject
 ///
@@ -4875,7 +4879,7 @@ void k_colorpicker__kcolorpicker_on_input_method_event(void* self, void (*callba
 /// @param self kColorPicker__KColorPicker*
 /// @param param1 enum Qt__InputMethodQuery
 ///
-QVariant* k_colorpicker__kcolorpicker_input_method_query(void* self, int64_t param1);
+QVariant* k_colorpicker__kcolorpicker_input_method_query(void* self, int32_t param1);
 
 /// Inherited from QWidget
 ///
@@ -4886,7 +4890,7 @@ QVariant* k_colorpicker__kcolorpicker_input_method_query(void* self, int64_t par
 /// @param self kColorPicker__KColorPicker*
 /// @param param1 enum Qt__InputMethodQuery
 ///
-QVariant* k_colorpicker__kcolorpicker_qbase_input_method_query(void* self, int64_t param1);
+QVariant* k_colorpicker__kcolorpicker_qbase_input_method_query(void* self, int32_t param1);
 
 /// Inherited from QWidget
 ///
@@ -4897,7 +4901,7 @@ QVariant* k_colorpicker__kcolorpicker_qbase_input_method_query(void* self, int64
 /// @param self kColorPicker__KColorPicker*
 /// @param callback QVariant* func(kColorPicker__KColorPicker* self, enum Qt__InputMethodQuery param1)
 ///
-void k_colorpicker__kcolorpicker_on_input_method_query(void* self, QVariant* (*callback)(void*, int64_t));
+void k_colorpicker__kcolorpicker_on_input_method_query(void* self, QVariant* (*callback)(void*, int32_t));
 
 /// Inherited from QWidget
 ///
