+++ include/extras-kcompletion/libkcombobox.h
@@ -996,7 +996,7 @@ void k_combobox_set_view(void* self, void* itemView);
 /// @param query enum Qt__InputMethodQuery
 /// @param argument QVariant*
 ///
-QVariant* k_combobox_input_method_query2(void* self, int64_t query, void* argument);
+QVariant* k_combobox_input_method_query2(void* self, int32_t query, void* argument);
 
 /// Inherited from QComboBox
 ///
@@ -2100,7 +2100,7 @@ void k_combobox_set_graphics_effect(void* self, void* effect);
 /// @param self KComboBox*
 /// @param type enum Qt__GestureType
 ///
-void k_combobox_grab_gesture(void* self, int64_t type);
+void k_combobox_grab_gesture(void* self, int32_t type);
 
 /// Inherited from QWidget
 ///
@@ -2109,7 +2109,7 @@ void k_combobox_grab_gesture(void* self, int64_t type);
 /// @param self KComboBox*
 /// @param type enum Qt__GestureType
 ///
-void k_combobox_ungrab_gesture(void* self, int64_t type);
+void k_combobox_ungrab_gesture(void* self, int32_t type);
 
 /// Inherited from QWidget
 ///
@@ -3098,7 +3098,7 @@ void k_combobox_set_parent(void* self, void* parent);
 /// @param parent QWidget*
 /// @param f flag of enum Qt__WindowType
 ///
-void k_combobox_set_parent2(void* self, void* parent, int64_t f);
+void k_combobox_set_parent2(void* self, void* parent, int32_t f);
 
 /// Inherited from QWidget
 ///
@@ -3176,9 +3176,9 @@ void k_combobox_add_action(void* self, void* action);
 /// [Upstream resources](https://doc.qt.io/qt-6/qwidget.html#addActions)
 ///
 /// @param self KComboBox*
-/// @param actions libqt_list /* of QAction* */
+/// @param actions libqt_list of QAction*
 ///
-void k_combobox_add_actions(void* self, libqt_list actions);
+void k_combobox_add_actions(void* self, libqt_list /* of QAction* */ actions);
 
 /// Inherited from QWidget
 ///
@@ -3186,9 +3186,9 @@ void k_combobox_add_actions(void* self, libqt_list actions);
 ///
 /// @param self KComboBox*
 /// @param before QAction*
-/// @param actions libqt_list /* of QAction* */
+/// @param actions libqt_list of QAction*
 ///
-void k_combobox_insert_actions(void* self, void* before, libqt_list actions);
+void k_combobox_insert_actions(void* self, void* before, libqt_list /* of QAction* */ actions);
 
 /// Inherited from QWidget
 ///
@@ -3215,7 +3215,9 @@ void k_combobox_remove_action(void* self, void* action);
 ///
 /// @param self KComboBox*
 ///
-libqt_list /* of QAction* */ k_combobox_actions(void* self);
+/// @return libqt_list of QAction*
+///
+libqt_list k_combobox_actions(void* self);
 
 /// Inherited from QWidget
 ///
@@ -3272,7 +3274,7 @@ QWidget* k_combobox_parent_widget(void* self);
 /// @param self KComboBox*
 /// @param type flag of enum Qt__WindowType
 ///
-void k_combobox_set_window_flags(void* self, int64_t type);
+void k_combobox_set_window_flags(void* self, int32_t type);
 
 /// Inherited from QWidget
 ///
@@ -3282,7 +3284,7 @@ void k_combobox_set_window_flags(void* self, int64_t type);
 ///
 /// @return flag of enum Qt__WindowType
 ///
-int64_t k_combobox_window_flags(void* self);
+int32_t k_combobox_window_flags(void* self);
 
 /// Inherited from QWidget
 ///
@@ -3291,7 +3293,7 @@ int64_t k_combobox_window_flags(void* self);
 /// @param self KComboBox*
 /// @param param1 enum Qt__WindowType
 ///
-void k_combobox_set_window_flag(void* self, int64_t param1);
+void k_combobox_set_window_flag(void* self, int32_t param1);
 
 /// Inherited from QWidget
 ///
@@ -3300,7 +3302,7 @@ void k_combobox_set_window_flag(void* self, int64_t param1);
 /// @param self KComboBox*
 /// @param type flag of enum Qt__WindowType
 ///
-void k_combobox_override_window_flags(void* self, int64_t type);
+void k_combobox_override_window_flags(void* self, int32_t type);
 
 /// Inherited from QWidget
 ///
@@ -3310,7 +3312,7 @@ void k_combobox_override_window_flags(void* self, int64_t type);
 ///
 /// @return enum Qt__WindowType
 ///
-int64_t k_combobox_window_type(void* self);
+int32_t k_combobox_window_type(void* self);
 
 /// Inherited from QWidget
 ///
@@ -3521,7 +3523,7 @@ void k_combobox_on_custom_context_menu_requested(void* self, void (*callback)(vo
 ///
 /// @return flag of enum Qt__InputMethodHint
 ///
-int64_t k_combobox_input_method_hints(void* self);
+int32_t k_combobox_input_method_hints(void* self);
 
 /// Inherited from QWidget
 ///
@@ -3530,7 +3532,7 @@ int64_t k_combobox_input_method_hints(void* self);
 /// @param self KComboBox*
 /// @param hints flag of enum Qt__InputMethodHint
 ///
-void k_combobox_set_input_method_hints(void* self, int64_t hints);
+void k_combobox_set_input_method_hints(void* self, int32_t hints);
 
 /// Inherited from QWidget
 ///
@@ -3615,7 +3617,7 @@ QPixmap* k_combobox_grab1(void* self, void* rectangle);
 /// @param type enum Qt__GestureType
 /// @param flags flag of enum Qt__GestureFlag
 ///
-void k_combobox_grab_gesture2(void* self, int64_t type, int32_t flags);
+void k_combobox_grab_gesture2(void* self, int32_t type, int32_t flags);
 
 /// Inherited from QWidget
 ///
@@ -3655,7 +3657,7 @@ void k_combobox_set_shortcut_auto_repeat2(void* self, int id, bool enable);
 /// @param param1 enum Qt__WindowType
 /// @param on bool
 ///
-void k_combobox_set_window_flag2(void* self, int64_t param1, bool on);
+void k_combobox_set_window_flag2(void* self, int32_t param1, bool on);
 
 /// Inherited from QWidget
 ///
@@ -3684,7 +3686,7 @@ QWidget* k_combobox_create_window_container2(void* window, void* parent);
 /// @param parent QWidget*
 /// @param flags flag of enum Qt__WindowType
 ///
-QWidget* k_combobox_create_window_container3(void* window, void* parent, int64_t flags);
+QWidget* k_combobox_create_window_container3(void* window, void* parent, int32_t flags);
 
 /// Inherited from QObject
 ///
@@ -3796,7 +3798,9 @@ void k_combobox_kill_timer2(void* self, int32_t id);
 ///
 /// @param self KComboBox*
 ///
-libqt_list /* of QObject* */ k_combobox_children(void* self);
+/// @return libqt_list of QObject*
+///
+libqt_list k_combobox_children(void* self);
 
 /// Inherited from QObject
 ///
@@ -4200,9 +4204,9 @@ int32_t k_combobox_completion_mode(void* self);
 ///
 /// @param self KComboBox*
 /// @param item enum KCompletionBase__KeyBindingType
-/// @param key libqt_list /* of QKeySequence* */
+/// @param key libqt_list of QKeySequence*
 ///
-bool k_combobox_set_key_binding(void* self, int32_t item, libqt_list key);
+bool k_combobox_set_key_binding(void* self, int32_t item, libqt_list /* of QKeySequence* */ key);
 
 /// Inherited from KCompletionBase
 ///
@@ -4211,7 +4215,9 @@ bool k_combobox_set_key_binding(void* self, int32_t item, libqt_list key);
 /// @param self KComboBox*
 /// @param item enum KCompletionBase__KeyBindingType
 ///
-libqt_list /* of QKeySequence* */ k_combobox_key_binding(void* self, int32_t item);
+/// @return libqt_list of QKeySequence*
+///
+libqt_list k_combobox_key_binding(void* self, int32_t item);
 
 /// Inherited from KCompletionBase
 ///
@@ -4406,7 +4412,7 @@ void k_combobox_on_event(void* self, bool (*callback)(void*, void*));
 /// @param self KComboBox*
 /// @param param1 enum Qt__InputMethodQuery
 ///
-QVariant* k_combobox_input_method_query(void* self, int64_t param1);
+QVariant* k_combobox_input_method_query(void* self, int32_t param1);
 
 /// Inherited from QComboBox
 ///
@@ -4417,7 +4423,7 @@ QVariant* k_combobox_input_method_query(void* self, int64_t param1);
 /// @param self KComboBox*
 /// @param param1 enum Qt__InputMethodQuery
 ///
-QVariant* k_combobox_qbase_input_method_query(void* self, int64_t param1);
+QVariant* k_combobox_qbase_input_method_query(void* self, int32_t param1);
 
 /// Inherited from QComboBox
 ///
@@ -4428,7 +4434,7 @@ QVariant* k_combobox_qbase_input_method_query(void* self, int64_t param1);
 /// @param self KComboBox*
 /// @param callback QVariant* func(KComboBox* self, enum Qt__InputMethodQuery param1)
 ///
-void k_combobox_on_input_method_query(void* self, QVariant* (*callback)(void*, int64_t));
+void k_combobox_on_input_method_query(void* self, QVariant* (*callback)(void*, int32_t));
 
 /// Inherited from QComboBox
 ///
@@ -6340,7 +6346,8 @@ void k_combobox_on_get_decoded_metric_f(void* self, double (*callback)(void*, in
 ///
 /// @warning Caller is responsible for freeing the returned memory using a similar sequence to:
 /// ```c
-/// // Example for freeing the returned map
+/// // Example for freeing the returned map of type:
+/// // libqt_map of enum KCompletionBase__KeyBindingType to libqt_list of QKeySequence*
 /// for (size_t i = 0; i < map.len; ++i) {
 ///     for (size_t j = 0; ((QKeySequence**)map.values)[i][j] != NULL; j++) {
 ///         free(((QKeySequence**)map.values)[i][j]);
@@ -6355,7 +6362,9 @@ void k_combobox_on_get_decoded_metric_f(void* self, double (*callback)(void*, in
 ///
 /// @param self KComboBox*
 ///
-libqt_map /* of enum KCompletionBase__KeyBindingType to libqt_list of QKeySequence* */ k_combobox_key_binding_map(void* self);
+/// @return libqt_map of enum KCompletionBase__KeyBindingType to libqt_list of QKeySequence*
+///
+libqt_map k_combobox_key_binding_map(void* self);
 
 /// Inherited from KCompletionBase
 ///
@@ -6363,7 +6372,8 @@ libqt_map /* of enum KCompletionBase__KeyBindingType to libqt_list of QKeySequen
 ///
 /// @warning Caller is responsible for freeing the returned memory using a similar sequence to:
 /// ```c
-/// // Example for freeing the returned map
+/// // Example for freeing the returned map of type:
+/// // libqt_map of enum KCompletionBase__KeyBindingType to libqt_list of QKeySequence*
 /// for (size_t i = 0; i < map.len; ++i) {
 ///     for (size_t j = 0; ((QKeySequence**)map.values)[i][j] != NULL; j++) {
 ///         free(((QKeySequence**)map.values)[i][j]);
@@ -6378,7 +6388,9 @@ libqt_map /* of enum KCompletionBase__KeyBindingType to libqt_list of QKeySequen
 ///
 /// @param self KComboBox*
 ///
-libqt_map /* of enum KCompletionBase__KeyBindingType to libqt_list of QKeySequence* */ k_combobox_qbase_key_binding_map(void* self);
+/// @return libqt_map of enum KCompletionBase__KeyBindingType to libqt_list of QKeySequence*
+///
+libqt_map k_combobox_qbase_key_binding_map(void* self);
 
 /// Inherited from KCompletionBase
 ///
@@ -6387,9 +6399,9 @@ libqt_map /* of enum KCompletionBase__KeyBindingType to libqt_list of QKeySequen
 /// Wrapper to allow overriding base class virtual or protected method
 ///
 /// @param self KComboBox*
-/// @param callback libqt_map /* of enum KCompletionBase__KeyBindingType to libqt_list of QKeySequence* */ func()
+/// @param callback libqt_map of enum KCompletionBase__KeyBindingType to libqt_list of QKeySequence* func()
 ///
-void k_combobox_on_key_binding_map(void* self, libqt_map /* of enum KCompletionBase__KeyBindingType to libqt_list of QKeySequence* */ (*callback)());
+void k_combobox_on_key_binding_map(void* self, libqt_map (*callback)());
 
 /// Inherited from KCompletionBase
 ///
@@ -6398,9 +6410,9 @@ void k_combobox_on_key_binding_map(void* self, libqt_map /* of enum KCompletionB
 /// Wrapper to allow calling virtual or protected method
 ///
 /// @param self KComboBox*
-/// @param keyBindingMap libqt_map /* of enum KCompletionBase__KeyBindingType to QKeySequence* */
+/// @param keyBindingMap libqt_map of enum KCompletionBase__KeyBindingType to QKeySequence*
 ///
-void k_combobox_set_key_binding_map(void* self, libqt_map /* of enum KCompletionBase__KeyBindingType to QKeySequence** */ keyBindingMap);
+void k_combobox_set_key_binding_map(void* self, libqt_map keyBindingMap);
 
 /// Inherited from KCompletionBase
 ///
@@ -6409,9 +6421,9 @@ void k_combobox_set_key_binding_map(void* self, libqt_map /* of enum KCompletion
 /// Wrapper to allow calling base class virtual or protected method
 ///
 /// @param self KComboBox*
-/// @param keyBindingMap libqt_map /* of enum KCompletionBase__KeyBindingType to QKeySequence* */
+/// @param keyBindingMap libqt_map of enum KCompletionBase__KeyBindingType to QKeySequence*
 ///
-void k_combobox_qbase_set_key_binding_map(void* self, libqt_map /* of enum KCompletionBase__KeyBindingType to QKeySequence** */ keyBindingMap);
+void k_combobox_qbase_set_key_binding_map(void* self, libqt_map keyBindingMap);
 
 /// Inherited from KCompletionBase
 ///
@@ -6420,9 +6432,9 @@ void k_combobox_qbase_set_key_binding_map(void* self, libqt_map /* of enum KComp
 /// Wrapper to allow overriding base class virtual or protected method
 ///
 /// @param self KComboBox*
-/// @param callback void func(KComboBox* self, libqt_map /* of enum KCompletionBase__KeyBindingType to QKeySequence* */ /* of enum KCompletionBase__KeyBindingType to QKeySequence* */)
+/// @param callback void func(KComboBox* self, libqt_map of enum KCompletionBase__KeyBindingType to QKeySequence*)
 ///
-void k_combobox_on_set_key_binding_map(void* self, void (*callback)(void*, libqt_map /* of enum KCompletionBase__KeyBindingType to QKeySequence** */));
+void k_combobox_on_set_key_binding_map(void* self, void (*callback)(void*, libqt_map));
 
 /// Inherited from KCompletionBase
 ///
