+++ src/libqnamespace.c
@@ -65,11 +65,11 @@ QKeyCombination* q_keycombination_new3() {
     return QKeyCombination_new3();
 }
 
-QKeyCombination* q_keycombination_new4(int64_t modifiers) {
+QKeyCombination* q_keycombination_new4(int32_t modifiers) {
     return QKeyCombination_new4(modifiers);
 }
 
-QKeyCombination* q_keycombination_new5(int64_t modifiers) {
+QKeyCombination* q_keycombination_new5(int32_t modifiers) {
     return QKeyCombination_new5(modifiers);
 }
 
@@ -81,11 +81,11 @@ QKeyCombination* q_keycombination_new7(int32_t key) {
     return QKeyCombination_new7(key);
 }
 
-QKeyCombination* q_keycombination_new8(int64_t modifiers, int32_t key) {
+QKeyCombination* q_keycombination_new8(int32_t modifiers, int32_t key) {
     return QKeyCombination_new8(modifiers, key);
 }
 
-QKeyCombination* q_keycombination_new9(int64_t modifiers, int32_t key) {
+QKeyCombination* q_keycombination_new9(int32_t modifiers, int32_t key) {
     return QKeyCombination_new9(modifiers, key);
 }
 
@@ -97,7 +97,7 @@ void q_keycombination_move_assign(void* self, void* other) {
     QKeyCombination_MoveAssign((QKeyCombination*)self, (QKeyCombination*)other);
 }
 
-int64_t q_keycombination_keyboard_modifiers(void* self) {
+int32_t q_keycombination_keyboard_modifiers(void* self) {
     return QKeyCombination_KeyboardModifiers((QKeyCombination*)self);
 }
